Move auto-run handler out of main.ts
